requirements.txt
html5lib==1.1
discord.py==1.5.1
requests==2.24.0
beautifulsoup4==4.9.3
discord==1.0.1
python-dotenv==0.15.0
